02/22/2017 2:36PM
* Condensed 3 player action scripts into 1.
* Updated the script execution order to help resolve some minor bugs.
* Deleted about 4 outdated, un-used scripts.
* Modified the machine gun script so it no longer needs a coroutine to fire shots.
* Modified the machine gun script so the crystal bullets now shoot out at the correct angle every single shot.

+++ Capstone Project/Assets/Scripts/Input/ControllableObject.cs
@@ -91,8 +91,8 @@ public class ControllableObject : MonoBehaviour {
     public int CurrentKey {
         get { return _currentKey; }
     }
-    private short _aimDirection;
-    public short AimDirection {
+    private int _aimDirection;
+    public int AimDirection {
         get { return _aimDirection; }
     }
